// packages/client/ui/src/components/organisms/data-table/types.ts
export type DataTableRowAction<TData> = {
  label: string
  onSelect: (row: TData) => void
  id?: string
  destructive?: boolean
  disabled?: (row: TData) => boolean
}
// ...
export type DataTableVariant = 'default' | 'content-manager'